refactor(admin): AdminSettings sous 300 lignes — éclatement complet (H3)
AdminSettings.vue passe de 633 à 103 lignes via le pattern G1 (orchestrateur
+ composable + sous-composants présentationnels).

- useAdminSettings.js : toute la logique (user, préférences localStorage,
  formulaire/validation mot de passe, déconnexion).
- Sous-composants admin : SettingsPersonalInfo, SettingsDisplaySection,
  SettingsNotifications (toggles v-model + émission save), SettingsSecurityCard,
  SettingsPermissions, SettingsSession, SettingsPasswordModal. CSS déplacé
  VERBATIM avec ses @media dans chaque sous-composant.

Parité : mêmes services (auth/toast/localStorage), même route /login, mêmes
classes CSS, mêmes textes/formats. Le $router.push devient useRouter().push.

Dette pré-existante conservée à l'identique :
- submitPasswordChange reste un stub (TODO API) : succès simulé, aucune requête.
- logout utilise confirm() natif ; console.log conservé dans savePreferences.

Tests : useAdminSettings (composable, services+router mockés), SettingsNotifications,
SettingsPasswordModal, SettingsPersonalInfo — 16/16 verts. Build vite exit 0.
